REFACTOR #38: Bestie 신청, 선택 API 관련 Controller, Service 코드 위치 변경 - Together Controller에 있던 API 요청들 ApplicantInfo Controller로 이동 - Together Service에 있던 비즈니스 로직들 ApplicantInfo Service로 이동
