+++ src/main/java/com/umc/FestieBE/domain/together/api/TogetherController.java
@@ -47,20 +47,4 @@ public class TogetherController {
         return ResponseEntity.ok().build();
     }
 
-    @PostMapping("/together/bestie/application")
-    public ResponseEntity<Void> createBestieApplication(
-            @Valid @RequestBody TogetherRequestDTO.BestieApplicationRequest request
-    ){
-        togetherService.createBestieApplication(request);
-        return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/together/bestie/choice")
-    public ResponseEntity<Void> createBestieChoice(
-            @Valid @RequestBody TogetherRequestDTO.BestieChoiceRequest request
-    ){
-        togetherService.createBestieChoice(request);
-        return ResponseEntity.ok().build();
-    }
-
 }
